+++ src/main/java/com/ssg/meowwms/service/warehouse/WarehouseService.java
@@ -1,6 +1,6 @@
 package com.ssg.meowwms.service.warehouse;
 
-import com.ssg.meowwms.dto.OptionDTO;
+import com.ssg.meowwms.dto.search.OptionDTO;
 import com.ssg.meowwms.dto.warehouse.WarehouseDTO;
 
 import java.util.List;
